Move cached_property to utils.py
