+++ sourmash/utils.py
@@ -1,3 +1,4 @@
+import functools
 import weakref
 
 from ._lowlevel import ffi, lib
@@ -76,3 +77,18 @@ def rustcall(func, *args):
     if backtrace:
         exc.rust_info = backtrace
     raise exc
+
+
+def cached_property(fun):
+    """A memoize decorator for class properties."""
+    @functools.wraps(fun)
+    def get(self):
+        try:
+            return self._cache[fun]
+        except AttributeError:
+            self._cache = {}
+        except KeyError:
+            pass
+        ret = self._cache[fun] = fun(self)
+        return ret
+    return property(get)
